pyproject.toml
[tool.poetry]
name = "todolite"
version = "0.1.0"
description = "Simple to-do list application"
authors = ["Pavel Filippi <pavel.filippi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.7"
django-taggit = "^3.1.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
mypy = "^0.991"
flake8 = "^6.0.0"
django-stubs = "^1.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = [
    "taggit.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "todolite.settings"
